dispatcher/requirements.txt
APScheduler==3.6.3
asgiref==3.2.7
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-filter==2.2.0
django-heroku==0.3.1
django-request==1.5.5
djangorestframework==3.11.0
djangorestframework-firebase==0.3.0
dnspython==1.16.0
firebase-admin==4.1.0
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
Markdown==3.2.1
msgpack==1.0.0
mysql-connector-python==8.0.19
mysqlclient==1.4.6
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.0.1
